requirements.txt
Flask>=3.0.0
PyMySQL>=1.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
pyinstaller>=6.0.0
